refactor(native): move runtime input into subdomain folder
