+++ src/native/engine/RESOURCE_COLLECTION_INTEGRATION_GUIDE.md
@@ -210,6 +210,6 @@ When connecting to TypeScript API layer:
 - **Tests**: `tests/native/player_resources_test.c` (28 tests)
 - **Tests**: `tests/native/interactive_collection_test.c` (26 tests)
 - **Tests**: `tests/native/interactive_object_spawner_test.c` (14 tests)
-- **Existing Systems**: `src/native/engine/runtime/input_binding.{h,c}`
+- **Existing Systems**: `src/native/engine/runtime/input/input_binding.{h,c}`
 - **Existing Systems**: `src/native/engine/runtime/terrain/terrain_chunks.{h,c}`
