+++ src/native/engine/runtime/interactive_collection.h
@@ -11,7 +11,7 @@
 #define BANANA_ENGINE_INTERACTIVE_COLLECTION_H
 
 #include "player/player_resources.h"
-#include "input_binding.h"
+#include "input/input_binding.h"
 
 #include <stdint.h>
