+++ src/native/engine/runtime/move_target_service.c
@@ -1,6 +1,6 @@
 #include "move_target_service.h"
 
-#include "input_contract.h"
+#include "input/input_contract.h"
 
 #include <math.h>
 #include <stdio.h>
