+++ tests/native/runtime/tick/runtime_tick_input_phase_test.c
@@ -1,5 +1,5 @@
 #include "runtime/tick/tick_input_phase.h"
-#include "runtime/input_contract.h"
+#include "runtime/input/input_contract.h"
 
 #include <stdio.h>
