+++ tests/native/runtime_move_target_domain_test.c
@@ -1,6 +1,6 @@
 #include "runtime/move_target_domain.h"
 #include "runtime/move_target_service.h"
-#include "runtime/input_contract.h"
+#include "runtime/input/input_contract.h"
 
 #include <stdio.h>
